:snowflake: Toggle wireframe implemented. Closes #10.
When 'a' is pressed, the the necessary objects switch their wireframe
state. It doesn't yet get updated on the screen, since the update cycle is
not yet implemented. Issue #12 has to be resolved before proceedin objects
switch their wireframe state. It doesn't yet get updated on the screen,
since the update cycle is not yet implemented. Issue #12 has to be
resolved before proceeding..