closes #39 #40. Bounding sphere fixes. Collisions now detected.
* Bounding spehere movement is now correct
* Collisions are now detected.

NOTE: When car touches a butter, its position is RESET. This is for
TESTING purposes (in correct version it should just stop). The reason I
set it as default, is because there is bug (issue #41). As soon as its
fixed, just stop the car and move it "backwards" if necessary.